+++ packages/wrapper-react/CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to npm module [@typefox/monaco-editor-react](https://www.npm
 
 ## [3.2.0] - 2024-04-12
 
-- Updated to `monaco-editor-wrapper@4.2.0`
+- Updated to `monaco-editor-wrapper@4.2.0` and `monaco-languageclient@8.3.0`
 - @typefox/monaco-editor-react works in strict mode [#634](https://github.com/TypeFox/monaco-languageclient/pull/634)
+- Using own `Logger` instance where implementation is provided by `monaco-languageclient`
 
 ## [3.1.0] - 2024-04-10